LPS-57497 move CSSBuilderUtil to css-builder module
